+++ bakefinity/src/main/java/com/springboot/bakefinity/utils/SecurityUtil.java
@@ -6,7 +6,7 @@ import org.springframework.security.core.context.SecurityContextHolder;
 
 public class SecurityUtil {
 
-    public static Long getCurrentUserId() {
+    public static Integer getCurrentUserId() {
         Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
         if (authentication instanceof JwtAuthentication) {
             return ((JwtAuthentication) authentication).getUserId();
